Add find items by category
