+++ gonggu-market-api/src/main/java/com/gonggu/market/api/domain/item/ItemRepository.java
@@ -2,5 +2,8 @@ package com.gonggu.market.api.domain.item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface ItemRepository extends JpaRepository<Item, Long> {
+    List<Item> findAllByCategory(String category);
 }
